Some panels require data to be sent as RBG rather than RGB
Some panels require data to be sent as RBG rather than RGB.

This PR adds a flag to control that (defaults for RGB as that is the current behaviour).  It adds a setter method for the flag and the required logic to fill the buffer in the appropriate order when the flag is set.

Usage:

Simply issue `display.setRBG(true);` in the sketch after `display.begin();`

If `display.setRBG() is not used or is set to false then the library functions exactly as before and fills the buffer for RGB.

The proposed change should not break any existing code.